requirements.txt
python-gitlab
python-dateutil
numpy
matplotlib
